add config based base64 decoding to kafka instrumentation
