+++ dd-java-agent/instrumentation/kafka-clients-0.11/kafka-clients-0.11.gradle
@@ -26,6 +26,7 @@ dependencies {
   testCompile group: 'javax.xml.bind', name: 'jaxb-api', version: '2.2.3'
   testCompile group: 'org.assertj', name: 'assertj-core', version: '2.9.+'
   testCompile group: 'org.mockito', name: 'mockito-core', version: '2.19.0'
+  testCompile deps.guava
 
   // Include latest version of kafka itself along with latest version of client libs.
   // This seems to help with jar compatibility hell.
@@ -36,4 +37,5 @@ dependencies {
   latestDepTestCompile group: 'org.springframework.kafka', name: 'spring-kafka', version: '2.2.+'
   latestDepTestCompile group: 'org.springframework.kafka', name: 'spring-kafka-test', version: '2.2.+'
   latestDepTestCompile group: 'org.assertj', name: 'assertj-core', version: '3.+'
+  latestDepTestCompile deps.guava
 }
