+++ dd-java-agent/instrumentation/kafka-clients-0.11/src/main/java/datadog/trace/instrumentation/kafka_clients/TextMapExtractAdapter.java
@@ -1,5 +1,6 @@
 package datadog.trace.instrumentation.kafka_clients;
 
+import datadog.trace.api.Config;
 import datadog.trace.bootstrap.instrumentation.api.AgentPropagation;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -9,7 +10,16 @@ import org.apache.kafka.common.header.Headers;
 
 public class TextMapExtractAdapter implements AgentPropagation.Getter<Headers> {
 
-  public static final TextMapExtractAdapter GETTER = new TextMapExtractAdapter();
+  public static final TextMapExtractAdapter GETTER =
+      new TextMapExtractAdapter(Config.get().isKafkaClientBase64DecodingEnabled());
+
+  private final boolean base64DecodeHeaders;
+  private final Base64Decoder base64;
+
+  public TextMapExtractAdapter(boolean base64DecodeHeaders) {
+    this.base64DecodeHeaders = base64DecodeHeaders;
+    this.base64 = base64DecodeHeaders ? new Base64Decoder() : null;
+  }
 
   @Override
   public Iterable<String> keys(final Headers headers) {
@@ -28,6 +38,9 @@ public class TextMapExtractAdapter implements AgentPropagation.Getter<Headers> {
     }
     if (header.value() == null) {
       return null;
+    }
+    if (base64DecodeHeaders) {
+      return new String(base64.decode(header.value()), StandardCharsets.UTF_8);
     }
     return new String(header.value(), StandardCharsets.UTF_8);
   }
