+++ dd-trace-api/dd-trace-api.gradle
@@ -9,8 +9,7 @@ excludedClassesCoverage += [
   'datadog.trace.api.DDTraceApiInfo',
   'datadog.trace.api.GlobalTracer*',
   'datadog.trace.api.CorrelationIdentifier',
-  'datadog.trace.api.DDTags',
-  'datadog.trace.api.DDId.Base64Decoder'
+  'datadog.trace.api.DDTags'
 ]
 
 description = 'dd-trace-api'
