+++ dd-trace-api/src/main/java/datadog/trace/api/config/TraceInstrumentationConfig.java
@@ -29,6 +29,8 @@ public final class TraceInstrumentationConfig {
   public static final String LOGS_INJECTION_ENABLED = "logs.injection";
 
   public static final String KAFKA_CLIENT_PROPAGATION_ENABLED = "kafka.client.propagation.enabled";
+  public static final String KAFKA_CLIENT_BASE64_DECODING_ENABLED =
+      "kafka.client.base64.decoding.enabled";
 
   private TraceInstrumentationConfig() {}
 }
